requirements.txt
certifi==2020.6.20
numpy==1.19.2
opencv-contrib-python==4.4.0.44
opencv-python==4.4.0.44
PyYAML==5.3.1
wincertstore==0.2
